Update ConsoleApplication1.cpp
I printed the code I have worked on so far. Thus far, I'm struggling to try to understand how the code is reading it. I'm testing it by printing phrases if it finds a piece and then checking the next piece diagonally downward from it.
The issue I'm having is that it doesn't seem to be reading it, instead, just skipping to the next line if it finds a piece.
Message me if you want to discuss my code so far